fix: api수정에 따른 basketService, basket page 수정

See merge request sw_track/class_03/web_project/team5/team5-commerce!99

+++ src/views/pages/basket/main.js
@@ -3,10 +3,9 @@ import {
   setLocalStorageItem,
   parsePrice,
 } from '/public/scripts/util.js';
-import cartService from '/public/scripts/cartService.js';
+import basketService from '/public/scripts/basketService.js';
 import productService from '/public/scripts/productService.js';
 
-// 전체 아이템 중 선택된 아이템만을 반환
 const selectedItems = () => {
   const items = document.querySelectorAll('.product');
   const selectedItems = [...items].filter(
@@ -15,7 +14,6 @@ const selectedItems = () => {
   return selectedItems;
 };
 
-// 변화 이벤트마다 실행하여 현재 선택된 품목에 의한 총 가격을 계산
 const printTotalPrice = () => {
   const productTotalPrice = document.querySelector('.product_total_price');
   const selected = selectedItems();
@@ -30,12 +28,10 @@ const printTotalPrice = () => {
   return;
 };
 
-// guset, user 공통 장바구니 RUD
-//
 const getLocalBasket = () => {
   return getLocalStorageItem('basket');
 };
-// 수량 변경 시 사용
+
 const setLocalBasket = (item) => {
   const productId = item.dataset.productId;
   const quantity = Number(item.querySelector('.product_count').value);
@@ -51,7 +47,13 @@ const setLocalBasket = (item) => {
   setLocalStorageItem('basket', updatedBasket);
   return;
 };
-// 물품 삭제 시 사용
+const getUserBaskets = async () => {
+  const { baskets } = await basketService.getCurrentUserBaskets();
+  return baskets;
+};
+const upToDateUserBasksets = async (toUpdateObj) => {
+  await basketService.updateBasketItem(toUpdateObj);
+};
 const deleteLocalBasket = (item) => {
   const productId = item.dataset.productId;
   const localBasket = getLocalBasket();
@@ -62,35 +64,21 @@ const deleteLocalBasket = (item) => {
   return;
 };
 
-// user only 장바구니 RUD -> usnig api
-// user에 의한 접근 시 장바구니를 생성
-const getUserCartItems = async () => {
-  const { orderSheets } = await cartService.getCurrentUserCart();
-  return orderSheets;
-};
-// 장바구니 수량 변경 시 사용
-const patchUserCartItem = async (item) => {
+const updateUserBasketItem = async (item) => {
   const productId = item.dataset.productId;
   const quantity = Number(item.querySelector('.product_count').value);
-  const toUpdate = { productId, quantity };
-  const response = await cartService.updateCartItem(toUpdate);
+  const toUpdateObj = { productId, quantity };
+  const response = await basketService.updateBasketItem(toUpdateObj);
   setLocalBasket(item);
   return response;
 };
-// 물품 삭제 시 사용
-const deleteUserCartItem = async (item) => {
-  const productId = item.dataset.productId;
-  const response = await cartService.deleteCartItem({ productId });
+
+const deleteUserBasketItem = async (item) => {
+  const basketId = item.dataset.basketId;
+  const response = await basketService.deleteBasketItem(basketId);
   return response;
 };
-// updateBasket -> loaclBasket과 db Basket을 합친 basket을 return
-const updateBasket = async (localBasket) => {
-  localBasket.forEach(async (item) => await cartService.updateCartItem(item));
-  const basket = await getUserCartItems();
-  return basket;
-};
 
-// guest, user 공통 주문목록 설정 -> payment page로 전달될 array
 const setOrderList = () => {
   const selected = selectedItems();
   const orderItems = [];
@@ -106,36 +94,39 @@ const getOrderList = () => {
   return getLocalStorageItem('orderList');
 };
 
-// 장바구니 item 정보를 이용하여 동적으로 생성한다
-const createItemWrapper = ({ _id, title, price, imageKey }, count) => {
+const createItemWrapper = (
+  basketId,
+  { _id: productId, title, price, imageKey },
+  count,
+) => {
   const itemWrapper = document.createElement('div');
   itemWrapper.className = 'flex-justify-between product';
-  itemWrapper.dataset.productId = _id;
+  itemWrapper.dataset.basketId = basketId;
+  itemWrapper.dataset.productId = productId;
   itemWrapper.innerHTML = `
-        <input type="checkbox" />
-        <div class="img_wrap">
-            <img src="${imageKey}" class="product_img" alt="${title}" />
-        </div>
-        <div class="product_name">${title}</div>
-        <div class="product_price">${price.toLocaleString('ko-kr')}</div>
-        <div>
-        <input type="number" class="product_count" value="${Number(
-          count,
-        )}" min="0" />
-        </div>
-  `;
+          <input type="checkbox" />
+          <div class="img_wrap">
+              <img src="${imageKey}" class="product_img" alt="${title}" />
+          </div>
+          <div class="product_name">${title}</div>
+          <div class="product_price">${price.toLocaleString('ko-kr')}</div>
+          <div>
+          <input type="number" class="product_count" value="${Number(
+            count,
+          )}" min="0" />
+          </div>
+    `;
   return itemWrapper;
 };
 
-const paintBasketItem = async ({ productId, quantity }) => {
+const paintBasketItem = async ({ _id: basketId = '', productId, quantity }) => {
   const basketItemList = document.querySelector('.basket_item_list');
   const basketItem = await productService.getProductByProductId(productId);
-  const itemWrapper = createItemWrapper(basketItem, quantity);
+  const itemWrapper = createItemWrapper(basketId, basketItem, quantity);
   basketItemList.appendChild(itemWrapper);
   return;
 };
 
-// eventHandlers 분리
 const basketClickEventHandler = async (e) => {
   if (e.target.id === 'basketChkAll') {
     const checkBoxes = document.querySelectorAll('input[type="checkbox"]');
@@ -143,25 +134,27 @@ const basketClickEventHandler = async (e) => {
     setOrderList();
   }
   if (e.target.classList.contains('remove_all_items')) {
-    const items = document.querySelectorAll('.product');
-    [...items]?.forEach(async (item) => {
+    if (confirm('모두 삭제하시겠습니까?')) {
       if (sessionStorage.getItem('token')) {
-        await deleteUserCartItem(item);
+        basketService.deleteUserBaskets();
       }
-      item.remove();
-    });
-    localStorage.removeItem('basket');
-    localStorage.removeItem('orderList');
+      localStorage.removeItem('basket');
+      localStorage.removeItem('orderList');
+      const items = document.querySelectorAll('.product');
+      items.forEach((item) => item.remove());
+    }
   }
   if (e.target.classList.contains('remove_selected_item')) {
-    const selected = selectedItems();
-    selected.forEach(async (item) => {
-      if (sessionStorage.getItem('token')) {
-        await deleteUserCartItem(item);
-      }
-      deleteLocalBasket(item);
-      item.remove();
-    });
+    if (confirm('선택 항목을 삭제하시겠습니까?')) {
+      const selected = selectedItems();
+      selected.forEach(async (item) => {
+        if (sessionStorage.getItem('token')) {
+          await deleteUserBasketItem(item);
+        }
+        deleteLocalBasket(item);
+        item.remove();
+      });
+    }
   }
   if (e.target.classList.contains('payment')) {
     if (!sessionStorage.getItem('token')) {
@@ -179,6 +172,7 @@ const basketClickEventHandler = async (e) => {
       }
     }
   }
+  setOrderList();
   printTotalPrice();
 };
 
@@ -186,7 +180,7 @@ const basketChangeEventHandler = async (e) => {
   if (e.target.classList.contains('product_count')) {
     const item = e.target.parentNode.parentNode;
     if (sessionStorage.getItem('token')) {
-      await patchUserCartItem(item);
+      await updateUserBasketItem(item);
     }
     setLocalBasket(item);
   }
@@ -194,11 +188,11 @@ const basketChangeEventHandler = async (e) => {
   printTotalPrice();
 };
 
-// 페이지 로딩 시 실행될 init function
 const init = async () => {
   const loginToken = sessionStorage.getItem('token') ?? null;
 
-  let localBasket = getLocalBasket() ?? [];
+  const localBasket = getLocalBasket() ?? [];
+  let basketItems = [];
 
   const basket = document.querySelector('.basket');
   const basketItemList = document.querySelector('.basket_item_list');
@@ -206,29 +200,14 @@ const init = async () => {
     <div class="flex-justify-between">장바구니에 상품이 없습니다</div>
   `;
 
-  localBasket = localBasket.map((item) => {
-    if (item.itemId) {
-      return { productId: item.itemId, quantity: item.count };
-    }
-    return item;
-  });
-
-  setLocalStorageItem('basket', localBasket);
-  let basketItems = localBasket;
-  basketItemList.innerHTML = '';
-
+  basketItems = localBasket;
   if (loginToken) {
-    basketItems = await updateBasket(localBasket);
-    basketItemList.innerHTML = '';
-    basketItems.forEach(paintBasketItem);
-  } else {
-    basketItems?.forEach(paintBasketItem);
+    localBasket.forEach(upToDateUserBasksets);
+    basketItems = await getUserBaskets();
   }
+  setTimeout(() => basketItems?.forEach(paintBasketItem), 500);
 
-  // 클릭 이벤트 - checkbox 개별/전체 선택/해제 // 삭제
   basket.addEventListener('click', basketClickEventHandler);
-
-  // change 이벤트 - 수량 등
   basket.addEventListener('change', basketChangeEventHandler);
 };
